BRAT Experiment [Resolves #136]
Functionality to manage a BRAT experiment using a named job posting sample.
Adds a new class, called BratExperiment. This is meant to encompass the following management tasks:

- Creating empty annotation files for all job postings in the sample, in the format that BRAT can read
- Dividing up job postings into allocations based on configuration
- Adding new users and initial allocations
- Adding new allocations to existing users
- Storing all relevant metadata on s3
- Parsing the results of user annotations into labels

This pull request just implements the tasks around the starting and management of experiments. The tasks around parsing the results are deferred to be implemented in a future pull request.

It is not meant to cover the installation of BRAT, or any other devops/infrastructure tasks, which would be out of scope of this repository. Installations are meant to sync this data from S3 on their own.